MenuPage aangemaakt en content 'MainWindow' verplaats naar 'MenuPage'. 'MainWindow' host gemaakt van 'MenuPage'. Pagina navigatie geïmplementeerd.
