+++ MenuPage.xaml
@@ -1,0 +1,60 @@
+﻿<Page x:Class="Slime_Busters_Menu.MenuPage"
+        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
+        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
+        xmlns:local="clr-namespace:Slime_Busters_Menu"
+        mc:Ignorable="d"
+        Title="MenuPage" Height="1080" Width="1920">
+
+    <Grid Height="1080px" Width="1920px">
+        <!-- background image -->
+        <Image Source="/main menu scenery.jpg" Stretch="Fill"/>
+
+        <!-- grid for buttons -->
+        <Grid>
+            <Grid.RowDefinitions>
+                <RowDefinition Height="368*"/>
+                <RowDefinition Height="250*"/>
+                <RowDefinition Height="250*"/>
+                <RowDefinition Height="250*"/>
+                <RowDefinition Height="368*"/>
+            </Grid.RowDefinitions>
+            <Grid.ColumnDefinitions>
+                <ColumnDefinition Width="734*"/>
+                <ColumnDefinition Width="452*"/>
+                <ColumnDefinition Width="734*"/>
+            </Grid.ColumnDefinitions>
+
+            <!-- Button Style -->
+            <Grid.Resources>
+                <Style TargetType="Button">
+                    <Setter Property="Background" Value="Transparent"/>
+                    <Setter Property="BorderBrush" Value="Transparent"/>
+                    <Setter Property="Template">
+                        <Setter.Value>
+                            <ControlTemplate TargetType="Button">
+                                <ContentPresenter/>
+                            </ControlTemplate>
+                        </Setter.Value>
+                    </Setter>
+                </Style>
+            </Grid.Resources>
+
+            <!-- buttons -->
+            <Button x:Name="Spelen" Grid.Row="1" Grid.Column="1" Background="Transparent" BorderBrush="Transparent" Click="SpelenButtonClick">
+                <Image Source="/Spelen Zwart.png"/>
+            </Button>
+
+            <Button x:Name="Winkel" Grid.Row="2" Grid.Column="1" Background="Transparent" BorderBrush="Transparent" Click="WinkelButtonClick">
+                <Image Source="/Winkel Zwart.png"/>
+            </Button>
+
+            <Button x:Name="Stoppen" Grid.Row="3" Grid.Column="1" Background="Transparent" BorderBrush="Transparent" Click="StoppenButtonClick">
+                <Image Source="/Stoppen.png"/>
+            </Button>
+            <ContentControl Name="MainContentControl"/>
+        </Grid>
+        <Frame Name="Menu" NavigationUIVisibility="Hidden" Height="1080px" Width="1920px"></Frame>
+    </Grid>
+</Page>
