+++ WinkelPage.xaml
@@ -11,6 +11,7 @@
     <Grid Height="1080px" Width="1920px">
         <Rectangle Width="1920" Height="1080" Fill="Blue" />
         <TextBlock Text="Winkel Page" FontSize="100" HorizontalAlignment="Center" VerticalAlignment="Center"/>
-
+        <Button x:Name="BackButton" Content="Back" HorizontalAlignment="Left" VerticalAlignment="Top" Margin="10" Click="BackButtonClick"/>
+        <Frame Name="Winkel" NavigationUIVisibility="Hidden" Height="1080px" Width="1920px"></Frame>
     </Grid>
 </Page>
